Initial Commit
In college I developed a fascination with the bounds of my very simple
understanding of math and image generation. Spurred on by a professor, I
created several iterations on a theme.

Honestly, I remember none of the details, but I often think of the
images. So here they will sit.